requirements.txt
dlib
lpips
numpy
scipy
torch
pillow
matplotlib
torchvision
opencv-python
python-telegram-bot